[OneBot] [Breaking Change] Completely change type of MessageId

+++ Lagrange.OneBot/Core/Entity/Action/Response/OneBotGetMessageResponse.cs
@@ -4,15 +4,15 @@ using Lagrange.OneBot.Core.Entity.Message;
 namespace Lagrange.OneBot.Core.Entity.Action.Response;
 
 [Serializable]
-public class OneBotGetMessageResponse(DateTime time, string messageType, uint messageId, List<OneBotSegment> message)
+public class OneBotGetMessageResponse(DateTime time, string messageType, int messageId, List<OneBotSegment> message)
 {
     [JsonPropertyName("time")] public int Time { get; set; } = (int)(time - DateTime.UnixEpoch).TotalSeconds;
     
     [JsonPropertyName("message_type")] public string MessageType { get; set; } = messageType;
 
-    [JsonPropertyName("message_id")] public int MessageId { get; set; } = (int)messageId;
+    [JsonPropertyName("message_id")] public int MessageId { get; set; } = messageId;
 
-    [JsonPropertyName("real_id")] public uint RealId { get; set; } = messageId;
+    [JsonPropertyName("real_id")] public int RealId { get; set; } = messageId;
     
     [JsonPropertyName("sender")] public OneBotSender Sender { get; set; } = new();
